fix: separated the login and signup to two different components, added loding while processing submit, added error handling
