+++ src/main.tsx
@@ -2,10 +2,10 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import './index.css' 
-import LoginPage from './pages/LoginPage'
+import AuthPage from './pages/AuthPage'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <LoginPage />
+    <AuthPage />
   </StrictMode>,
 )
